+++ package.json
@@ -55,7 +55,6 @@
     "react-flip-toolkit": "^7.1.0",
     "react-hook-form": "^7.45.4",
     "react-toastify": "^9.1.3",
-    "superjson": "^1.13.1",
     "tailwindcss": "^3.3.3",
     "unique-names-generator": "^4.7.1",
     "unplugin-icons": "^0.17.0",
