+++ src/components/Footer.tsx
@@ -1,4 +1,4 @@
-import { selectionAtom } from "@/pages/scan";
+import { selectionAtom } from "@/utils/scan";
 import { useAtomValue } from "jotai";
 import Link from "next/link";
 import { useRouter } from "next/router";
